Remove unnecessary property TransferFile
This property was only used for testing and unnecessarily exposed its
state to the outside world. This problem originated probably from the
internal methods of the Validator class, that also relied on this
property and changed its state.

To simplify testing and reduce complexity, the TransferFile property
no longer exists and the methods behave in a more "functional way"
by retrieving the transfer file as parameter and generating a
specific output if necessary.